Add `from_config` util method for generating framework-independent components from an arbitrary config (dict, json, yaml, file). Plus tests.
