--- rllib/components/__init__.py
@@ -1,8 +1,0 @@
-from ray.rllib.components.component import Component
-
-__all__ = [
-    "Component",
-]
-
-# TODO: Move all sharable components here (from their current locations).
-# TODO: e.g. Memories, Optimizers, ...
