--- rllib/components/tests/dummy_config.yml
@@ -1,5 +1,0 @@
-type: ray.rllib.components.tests.test_components.DummyComponent
-prop_a: something else
-prop_b: 1.0
-prop_d: 3
-framework: torch
