+++ rllib/utils/tests/dummy_config.json
@@ -1,5 +1,5 @@
 {
-  "type": "ray.rllib.components.tests.test_components.DummyComponent",
+  "type": "ray.rllib.utils.tests.test_framework_independent_components.DummyComponent",
   "prop_a": "some-value",
   "prop_b": 2.0,
   "some_other_kwarg": "this_is_a_component_from_json_file"
