+++ rllib/utils/tests/dummy_config.yml
@@ -1,0 +1,5 @@
+type: ray.rllib.utils.tests.test_framework_independent_components.DummyComponent
+prop_a: something else
+prop_b: 1.0
+prop_d: 3
+framework: torch
