+++ rllib/utils/tests/test_framework_independent_components.py
@@ -1,7 +1,7 @@
 from abc import ABCMeta, abstractmethod
 import unittest
 
-from ray.rllib.components.component import Component
+from ray.rllib.utils.from_config import from_config
 from ray.rllib.utils.test_utils import check
 from ray.rllib.utils.framework import try_import_tf, try_import_torch
 
@@ -11,50 +11,54 @@ tf.enable_eager_execution()
 torch, _ = try_import_torch()
 
 
-class TestComponents(unittest.TestCase):
+class TestFrameWorkIndependentComponents(unittest.TestCase):
     """
     Tests the Component base class to implement framework-agnostic functional
     units.
     """
-    def test_dummy_component(self):
+    def test_dummy_components(self):
         # Switch on eager for testing purposes.
         tf.enable_eager_execution()
 
-        # Try to create a Component from a class w/o default constructor.
+        # Try to create from an abstract class w/o default constructor.
         # Expect None.
-        test = AbstractDummyComponent.from_config({"framework": "torch"})
+        test = from_config(
+            {"type": AbstractDummyComponent, "framework": "torch"})
         check(test, None)
 
         # Create a Component via python API (config dict).
-        component = Component.from_config(dict(
+        component = from_config(dict(
             type=DummyComponent,
             prop_a=1.0, prop_d="non_default"
         ))
         check(component.prop_d, "non_default")
 
         # Create a tf Component from json file.
-        component = Component.from_config("dummy_config.json")
+        component = from_config("dummy_config.json")
         check(component.prop_c, "default")
         check(component.prop_d, 4)  # default
         check(component.add(3.3).numpy(), 5.3)  # prop_b == 2.0
 
         # Create a torch Component from yaml file.
-        component = Component.from_config("dummy_config.yml")
+        component = from_config("dummy_config.yml")
         check(component.prop_a, "something else")
         check(component.prop_d, 3)
         check(component.add(1.2), torch.Tensor([2.2]))  # prop_b == 1.0
 
         # Create tf Component from json-string (e.g. on command line).
-        component = DummyComponent.from_config(
-            '{"prop_a": "A", "prop_b": -1.0, "prop_c": "non-default"}'
+        component = from_config(
+            '{"type": "ray.rllib.utils.tests.'
+            'test_framework_independent_components.DummyComponent", '
+            '"prop_a": "A", "prop_b": -1.0, "prop_c": "non-default"}'
         )
         check(component.prop_a, "A")
         check(component.prop_d, 4)  # default
         check(component.add(-1.1).numpy(), -2.1)  # prop_b == -1.0
 
         # Create torch Component from yaml-string.
-        component = Component.from_config(
-            "type: ray.rllib.components.tests.test_components.DummyComponent\n"
+        component = from_config(
+            "type: ray.rllib.utils.tests."
+            "test_framework_independent_components.DummyComponent\n"
             "prop_a: B\nprop_b: -1.5\nprop_c: non-default\nframework: torch"
         )
         check(component.prop_a, "B")
@@ -62,7 +66,7 @@ class TestComponents(unittest.TestCase):
         check(component.add(-5.1), torch.Tensor([-6.6]))  # prop_b == -1.5
 
 
-class DummyComponent(Component, metaclass=ABCMeta):
+class DummyComponent:
     """
     A simple DummyComponent that can be used for testing framework-agnostic
     logic. Implements a simple `add()` method for adding a value to
@@ -71,7 +75,7 @@ class DummyComponent(Component, metaclass=ABCMeta):
     def __init__(
             self, prop_a, prop_b=0.5, prop_c=None, framework="tf", **kwargs
     ):
-        super().__init__(framework)
+        self.framework = framework
         self.prop_a = prop_a
         self.prop_b = prop_b
         self.prop_c = prop_c or "default"
